factor out `IsServerError` and instances to its own module
To allow the `Handler` specific code for computing `W.Tx` certificates
we factor out the `IsServerError` definition and all the instances for
the Wallet types. Instances defined in `Api.Types` are left at definition
site.
